serial buffer counters reduced to uint8_t

+++ libraries/SingleSerial/SingleSerial.h
@@ -136,8 +136,8 @@ public:
 	///
 	/// Public so the interrupt handlers can see it
 	struct Buffer {
-		volatile uint16_t head, tail;	///< head and tail pointers
-		uint16_t mask;					///< buffer size mask for pointer wrap
+		volatile uint8_t head, tail;	///< head and tail pointers
+		uint8_t mask;					///< buffer size mask for pointer wrap
 		uint8_t *bytes;					///< pointer to allocated buffer
 	};
